+++ src/mixins/labels/coordinate.ts
@@ -2,7 +2,7 @@ import * as THREE from "three";
 
 import { LABEL_TYPES } from "../../enums";
 import settings from "../../settings";
-import { getArrayFromVector } from "../threeJsUtils";
+import { getArrayFromVector } from "../utils_three";
 import { BaseLabelsManager } from "./base";
 
 // @ts-ignore
