+++ src/mixins/measurements/angle.ts
@@ -3,7 +3,7 @@ import * as THREE from "three";
 import { MEASUREMENT_MODES_ENUM } from "../../enums";
 import { AngleLabelsManager } from "../labels/angle";
 import { LabelsManagerConstructor } from "../labels/base";
-import { calculateAngleBetweenAtoms, calculateAngleLabelPosition } from "../threeJsUtils";
+import { calculateAngleBetweenAtoms, calculateAngleLabelPosition } from "../utils_three";
 import { BaseMeasurementManager } from "./base";
 
 export class AnglesMeasurementManager extends BaseMeasurementManager<AngleLabelsManager> {
